@@ -5,7 +5,7 @@
   "scripts": {
     "gulp": "gulp",
     "lint": "eslint src/**/*.js",
-    "start": "webpack-dev-server",
+    "start": "webpack serve",
     "build": "webpack",
     "audio": "gulp audio",
     "images": "gulp images",
@@ -22,30 +22,27 @@
   ],
   "author": "Matt Surabian",
   "license": "MIT",
+  "overrides": {
+    "fsevents": "^2.3.3",
+    "chokidar": "^3.5.3"
+  },
   "devDependencies": {
-    "audiosprite": "^0.6.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^6.2.10",
-    "babel-preset-es2015": "^6.22.0",
-    "bluebird": "^3.7.2",
-    "eslint": "^4.18.2",
-    "eslint-plugin-import": "^2.20.2",
-    "exports-loader": "^0.6.3",
-    "file-loader": "^0.10.0",
-    "glob": "^7.1.6",
-    "gsap": "^1.20.6",
+    "@babel/core": "^7.23.0",
+    "@babel/preset-env": "^7.23.0",
+    "audiosprite": "^0.7.2",
+    "babel-loader": "^9.1.3",
+    "eslint": "^8.50.0",
+    "eslint-plugin-import": "^2.29.0",
+    "glob": "^10.3.0",
+    "gsap": "^3.12.0",
     "gulp": "^4.0.2",
     "gulp-connect": "^5.7.0",
     "gulp-shell": "^0.8.0",
-    "hard-source-webpack-plugin": "^0.12.0",
-    "howler": "^2.1.3",
-    "json-loader": "^0.5.4",
+    "howler": "^2.2.4",
     "lodash": "^4.17.21",
-    "node-object-hash": "^1.4.2",
-    "pixi.js": "^4.8.9",
-    "script-loader": "^0.7.2",
-    "tween.js": "^16.6.0",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.9.1"
+    "pixi.js": "^7.4.2",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.0"
   }
 }
